requirements.txt
Flask
tensorflow
numpy
requests
Pillow
beautifulsoup4
